back/requirements.txt
flask
flask_cors
mysql-connector-python
bcrypt
PyJWT
gunicorn
Flask-CORS
pymysql